requirements.txt
pygame-ce>=2.5.0
tcod>=16.0.0
numpy>=1.26.0
pytest>=8.0.0
